backend/requirements.txt
PyYAML
pymupdf
Pillow
pandas
plotly
opencv-python-headless
numpy
fastapi
uvicorn[standard]
python-multipart
